Instead of scp'ing files around, switch to stdin
The scp approach is fragile: requires that the home
directory of the remote user is writable which may not be
true, and files that are unique locally may not be unique
remotely. So, the preferred method is now to use the '--file'
option for aii-shellfe. That option allows stdin ('-').
aii-installfe translates any and all file arguments (e.g.
the use of 'installlist') into a single '--file -' for
the remote/local aii-shellfe.