more correctly count actions
Previously, the program was aborting if there were no
command-line options specified, saying "no actions specified".
But just specifying --debug would be enough for it to think
that an action was being invoked, so it was bogus. Furthermore,
the new --file changes means that at the time of the check,
the options may not be filled in and so it might think that
there were no actions. Instead of this, I've made an explicit
action count which is based on hosts or files which are specified
(and not on arbitrary options).