Do not generate "part ..." lines for ignored disks
If "ignoredisk" is used to tell Anaconda to ignore a certain disk, then
it will not scan that disk for partitions, and any "part ..." lines
referring to such partitions will cause a "Partition not found" error.

Make sure such "part ..." lines are not emitted; handling such
partitions is now completely up to the %pre script and
ncm-filesystems/ncm-fstab.