apply "trust" values for ccm
The ccm code has been updated to allow for encrypted
profiles, which requires the client code to know whom to
trust. This data is stored within the profile (yes, this is
recursive). This patch simply reads out the profile value
and makes sure that we construct a valid /etc/ccm.conf during
kickstart (if we use the encrypted URL but don't have trust
configured, then the ccm-fetch will fail).