Fix dependecies in 13.1.1
Separate the OS dependencies of quattor client rpms from the rpms themselves. All the OS dependencies of the quattor client rpms are defined in AII_OSINSTALL_PACKAGES and the quattor client rpms in AII_OSINSTALL_PACKAGES. This ensures that all OS packages come from the base OS installation rather than a mix of the base OS and the current errata level, which causes all sorts of dependencies mess with at least perl. It requires support for the kickstart repo parameter for epel/rpmforge or similar packages such as perl-JSON-XS.