make wipe_metadata faster
Get the size of the drive in logical sectors (512 bytes), then wipe the first 10 and last 20 sectors.

We have to use fdisk because parted in SL5 will refuse to display the size of unlabeled drive.

The first 10 and last 20 sectors were chosen after some googling. They should be enough to wipe gpt partition information and raid metadata.