Fix update removes and change unittests
The update hook does not remove the VM images or the VM template.
The user must use remove hook to do that.
Due to this change the unittest was also updated to pass all the tests.
The aii will show a warning message if the user wants to reuse a old image.

Resolves: #199
See also: #204