Avoid CCM messing with the umask
If world_readable is not set in the CCM configuration, then the CCM
library modifies the umask to enforce the protection. However, CCM does
not restore the umask when it returns, and that causes problems for
modules which do not explicitly set permissions on newly created files
and/or libraries. Even common helpers like CAF::FileWriter have
problems. E.g., if you tell CAF::FileWriter to create "a/b/c" where
neither "a" nor "b" exists, then it will set the permissions on file "c"
and directory "b", but not on the top-level direcory "a".

Setting the world_writable flag disables the umask modifications in CCM,
so files/directories which do not have an explicit permission assigned
will be created according to the default umask inherited from the
environment.