aii-ks: Add a new hook for the "%post --nochroot" phase
This allows writing a plugin which opportunistically preserves content
(like SSH private host keys or Kerberos keytabs) by mounting the
previous filesystem(s) and copying files to the installer's ramdisk in
the pre_install phase, and restoring such saved files in the
post_install_nochroot phase.

Before RH7, it was possible to use the pre_install_end hook for
restoring such saved content to the newly formatted filesystem(s). But
Anaconda in RH7 wants to format the root filesystem itself, so restoring
the files now needs to happen later.