aii-pxelinux: Fix UEFI boot over http
If the kernel/initrd specified in the profile uses the
"(http,XXXX)/path/to/image" syntax, then the extra slash which is
currently inserted unconditionally at the beginning causes Grub to fail.
Make sure a path separator is inserted only if GRUB2_EFI_KERNEL_ROOT is
in fact not empty. Allow disabling the path mangling by setting
grub2_efi_kernel_root to 'none'.